refactor: Move printer to output package
