+++ cmd/oras/internal/display/status/deprecated.go
@@ -16,6 +16,7 @@ limitations under the License.
 package status
 
 import (
+	"oras.land/oras/cmd/oras/internal/output"
 	"sync"
 
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
@@ -28,7 +29,7 @@ import (
 
 // NewTagStatusHintPrinter creates a wrapper type for printing
 // tag status and hint.
-func NewTagStatusHintPrinter(printer *Printer, target oras.Target, refPrefix string) oras.Target {
+func NewTagStatusHintPrinter(printer *output.Printer, target oras.Target, refPrefix string) oras.Target {
 	var printHint sync.Once
 	var printHintErr error
 	onTagging := func(desc ocispec.Descriptor, tag string) error {
@@ -45,7 +46,7 @@ func NewTagStatusHintPrinter(printer *Printer, target oras.Target, refPrefix str
 }
 
 // NewTagStatusPrinter creates a wrapper type for printing tag status.
-func NewTagStatusPrinter(printer *Printer, target oras.Target) oras.Target {
+func NewTagStatusPrinter(printer *output.Printer, target oras.Target) oras.Target {
 	return listener.NewTagListener(target, nil, func(desc ocispec.Descriptor, tag string) error {
 		return printer.Println("Tagged", tag)
 	})
