+++ cmd/oras/root/tag.go
@@ -18,6 +18,7 @@ package root
 import (
 	"errors"
 	"fmt"
+	"oras.land/oras/cmd/oras/internal/output"
 
 	"github.com/spf13/cobra"
 	"oras.land/oras-go/v2"
@@ -98,7 +99,7 @@ Example - Tag the manifest 'v1.0.1' to 'v1.0.2' in an OCI image layout folder 'l
 
 func tagManifest(cmd *cobra.Command, opts *tagOptions) error {
 	ctx, logger := command.GetLogger(cmd, &opts.Common)
-	printer := status.NewPrinter(cmd.OutOrStdout())
+	printer := output.NewPrinter(cmd.OutOrStdout())
 	target, err := opts.NewTarget(opts.Common, logger)
 	if err != nil {
 		return err
